refactor: 유틸,검증 로직 폴더 분리
